agent: Add handler for `FillHashInvalidated` events
