+++ beamer/tests/agent/unit/test_handlers.py
@@ -1,131 +1,25 @@
 from copy import deepcopy
-from pathlib import Path
-from typing import Tuple
-from unittest.mock import MagicMock
 
 import pytest
-from eth_account import Account
-from eth_typing import BlockNumber
 from hexbytes import HexBytes
-from web3.types import BlockData, ChecksumAddress, Timestamp, Wei
+from web3.types import ChecksumAddress, Wei
 
-from beamer.chain import Context, claim_request, process_claims, process_requests
-from beamer.config import Config
-from beamer.events import ClaimMade, InitiateL1ResolutionEvent, RequestResolved
-from beamer.models.claim import Claim
-from beamer.models.request import Request
+from beamer.chain import claim_request, process_claims, process_requests
+from beamer.events import InitiateL1ResolutionEvent, RequestResolved
 from beamer.state_machine import process_event
+from beamer.tests.agent.unit.utils import (
+    ACCOUNT,
+    ADDRESS1,
+    CLAIM_ID,
+    REQUEST_ID,
+    TARGET_CHAIN_ID,
+    TIMESTAMP,
+    make_claim,
+    make_context,
+    make_request,
+)
 from beamer.tests.agent.utils import make_address
-from beamer.tracker import Tracker
-from beamer.typing import URL, ChainId, ClaimId, FillId, RequestId, Termination, TokenAmount
-from beamer.util import TokenMatchChecker
-
-SOURCE_CHAIN_ID = ChainId(2)
-TARGET_CHAIN_ID = ChainId(3)
-
-REQUEST_ID = RequestId(10)
-CLAIM_ID = ClaimId(200)
-FILL_ID = FillId(b"abc")
-
-CLAIMER_STAKE = Wei(10_000_000)
-CHALLENGER_STAKE = Wei(5_000_000)
-TERMINATION = Termination(1)
-TIMESTAMP = Timestamp(457)
-
-ACCOUNT = Account.from_key(0xB25C7DB31FEED9122727BF0939DC769A96564B2DE4C4726D035B36ECF1E5B364)
-
-ADDRESS1 = make_address()
-
-
-class MockEth:
-    def __init__(self, chain_id):
-        self.chain_id = chain_id
-
-    wait_for_transaction_receipt = MagicMock()
-
-
-class MockWeb3:
-    def __init__(self, chain_id):
-        self.eth = MockEth(chain_id=chain_id)
-
-
-def make_request() -> Request:
-    return Request(
-        request_id=REQUEST_ID,
-        source_chain_id=SOURCE_CHAIN_ID,
-        target_chain_id=TARGET_CHAIN_ID,
-        source_token_address=make_address(),
-        target_token_address=make_address(),
-        target_address=make_address(),
-        amount=TokenAmount(123),
-        valid_until=TIMESTAMP - 1,
-    )
-
-
-def make_claim(
-    request: Request,
-    claimer: ChecksumAddress = None,
-    claimer_stake: Wei = CLAIMER_STAKE,
-    challenger: ChecksumAddress = None,
-    challenger_stake: Wei = CHALLENGER_STAKE,
-    termination: Termination = TERMINATION,
-    fill_id: FillId = FILL_ID,
-) -> Claim:
-    challenger = challenger or make_address()
-    claim = Claim(
-        claim_made=ClaimMade(
-            chain_id=request.source_chain_id,
-            tx_hash=HexBytes(b""),
-            claim_id=CLAIM_ID,
-            request_id=request.id,
-            fill_id=fill_id,
-            claimer=claimer or make_address(),
-            claimer_stake=claimer_stake,
-            last_challenger=challenger,
-            challenger_stake_total=challenger_stake,
-            termination=termination,
-        ),
-        challenge_back_off_timestamp=123,
-    )
-    claim.add_challenger_stake(challenger, challenger_stake)
-    return claim
-
-
-def make_context() -> Tuple[Context, Config]:
-    checker = TokenMatchChecker([])
-    config = Config(
-        account=ACCOUNT,
-        deployment_info={},
-        l1_rpc_url=URL(""),
-        l2a_rpc_url=URL(""),
-        l2b_rpc_url=URL(""),
-        token_match_file=Path(),
-        fill_wait_time=1,
-        prometheus_metrics_port=None,
-    )
-
-    context = Context(
-        requests=Tracker(),
-        claims=Tracker(),
-        request_manager=MagicMock(),
-        fill_manager=MagicMock(),
-        match_checker=checker,
-        address=config.account.address,
-        latest_blocks={
-            SOURCE_CHAIN_ID: BlockData(
-                {
-                    "number": BlockNumber(42),
-                    "timestamp": TIMESTAMP,
-                }
-            )
-        },
-        config=config,
-        web3_l1=MagicMock(),
-        resolution_pool=MagicMock(),
-        l1_resolutions={},
-    )
-
-    return context, config
+from beamer.typing import Termination
 
 
 def test_skip_not_self_filled():
